genai-insight-bot/requirements.txt
python-telegram-bot==20.7
g4f==0.1.9
psycopg2-binary==2.9.7
supabase==1.0.3
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
aiohttp==3.8.5
matplotlib==3.7.2
networkx==3.1
plotly==5.15.0
python-multipart==0.0.6
sqlalchemy==2.0.20
alembic==1.12.0
pgvector==0.2.0